requirements.txt
bcrypt
fastapi[standard]
jose
passlib
psycopg2
python-jose
pydantic_settings
python-dotenv
SQLAlchemy
